Update Structure Server App
