--- server-app/src/main/java/com/app/ServerAppApplication.java
@@ -1,34 +1,0 @@
-package com.app;
-
-import java.util.Arrays;
-import java.util.Collections;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-import org.springframework.web.filter.CorsFilter;
-
-@SpringBootApplication
-public class ServerAppApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ServerAppApplication.class, args);
-	}
-	
-	
-	@Bean
-	public CorsFilter corsFilter() {
-	    final UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-
-	    final CorsConfiguration config = new CorsConfiguration();
-	    config.setAllowCredentials(true);
-	    config.setAllowedOrigins(Collections.singletonList("*"));
-	    config.setAllowedHeaders(Arrays.asList("Origin", "Content-Type", "Accept"));
-	    config.setAllowedMethods(Arrays.asList("GET", "POST", "PUT", "DELETE", "OPTIONS"));
-
-	    source.registerCorsConfiguration("/**", config);
-	    return new CorsFilter(source);
-	}
-}
