--- server-app/src/main/java/com/app/entity/Users.java
@@ -1,127 +1,0 @@
-package com.app.entity;
-
-import java.util.Date;
-
-import org.apache.ibatis.annotations.Mapper;
-import org.springframework.stereotype.Component;
-
-@Component
-@Mapper
-public class Users {
-
-	/**
-	 * This field was generated by MyBatis Generator. This field corresponds to the database column users.id
-	 * @mbg.generated  Fri Mar 04 00:33:19 ICT 2022
-	 */
-	private Integer id;
-	/**
-	 * This field was generated by MyBatis Generator. This field corresponds to the database column users.work_name
-	 * @mbg.generated  Fri Mar 04 00:33:19 ICT 2022
-	 */
-	private String workName;
-	/**
-	 * This field was generated by MyBatis Generator. This field corresponds to the database column users.starting_date
-	 * @mbg.generated  Fri Mar 04 00:33:19 ICT 2022
-	 */
-	private Date startingDate;
-	/**
-	 * This field was generated by MyBatis Generator. This field corresponds to the database column users.ending_date
-	 * @mbg.generated  Fri Mar 04 00:33:19 ICT 2022
-	 */
-	private Date endingDate;
-	/**
-	 * This field was generated by MyBatis Generator. This field corresponds to the database column users.status
-	 * @mbg.generated  Fri Mar 04 00:33:19 ICT 2022
-	 */
-	private Integer status;
-
-	/**
-	 * This method was generated by MyBatis Generator. This method returns the value of the database column users.id
-	 * @return  the value of users.id
-	 * @mbg.generated  Fri Mar 04 00:33:19 ICT 2022
-	 */
-	public Integer getId() {
-		return id;
-	}
-
-	/**
-	 * This method was generated by MyBatis Generator. This method sets the value of the database column users.id
-	 * @param id  the value for users.id
-	 * @mbg.generated  Fri Mar 04 00:33:19 ICT 2022
-	 */
-	public void setId(Integer id) {
-		this.id = id;
-	}
-
-	/**
-	 * This method was generated by MyBatis Generator. This method returns the value of the database column users.work_name
-	 * @return  the value of users.work_name
-	 * @mbg.generated  Fri Mar 04 00:33:19 ICT 2022
-	 */
-	public String getWorkName() {
-		return workName;
-	}
-
-	/**
-	 * This method was generated by MyBatis Generator. This method sets the value of the database column users.work_name
-	 * @param workName  the value for users.work_name
-	 * @mbg.generated  Fri Mar 04 00:33:19 ICT 2022
-	 */
-	public void setWorkName(String workName) {
-		this.workName = workName == null ? null : workName.trim();
-	}
-
-	/**
-	 * This method was generated by MyBatis Generator. This method returns the value of the database column users.starting_date
-	 * @return  the value of users.starting_date
-	 * @mbg.generated  Fri Mar 04 00:33:19 ICT 2022
-	 */
-	public Date getStartingDate() {
-		return startingDate;
-	}
-
-	/**
-	 * This method was generated by MyBatis Generator. This method sets the value of the database column users.starting_date
-	 * @param startingDate  the value for users.starting_date
-	 * @mbg.generated  Fri Mar 04 00:33:19 ICT 2022
-	 */
-	public void setStartingDate(Date startingDate) {
-		this.startingDate = startingDate;
-	}
-
-	/**
-	 * This method was generated by MyBatis Generator. This method returns the value of the database column users.ending_date
-	 * @return  the value of users.ending_date
-	 * @mbg.generated  Fri Mar 04 00:33:19 ICT 2022
-	 */
-	public Date getEndingDate() {
-		return endingDate;
-	}
-
-	/**
-	 * This method was generated by MyBatis Generator. This method sets the value of the database column users.ending_date
-	 * @param endingDate  the value for users.ending_date
-	 * @mbg.generated  Fri Mar 04 00:33:19 ICT 2022
-	 */
-	public void setEndingDate(Date endingDate) {
-		this.endingDate = endingDate;
-	}
-
-	/**
-	 * This method was generated by MyBatis Generator. This method returns the value of the database column users.status
-	 * @return  the value of users.status
-	 * @mbg.generated  Fri Mar 04 00:33:19 ICT 2022
-	 */
-	public Integer getStatus() {
-		return status;
-	}
-
-	/**
-	 * This method was generated by MyBatis Generator. This method sets the value of the database column users.status
-	 * @param status  the value for users.status
-	 * @mbg.generated  Fri Mar 04 00:33:19 ICT 2022
-	 */
-	public void setStatus(Integer status) {
-		this.status = status;
-	}
-}
