+++ server-app/src/main/java/demo/dc/app/controller/UsersController.java
@@ -1,0 +1,21 @@
+package demo.dc.app.controller;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
+
+import demo.dc.domain.service.UserService;
+
+@RestController
+public class UsersController {
+	@Autowired
+	UserService userService;
+	
+	@RequestMapping(value = "/api/user", method = RequestMethod.GET)
+	public ResponseEntity<Object> getProduct(){
+		return new ResponseEntity<Object>(userService.getUsers(), HttpStatus.OK);
+	}
+}
