--- server-app/src/test/java/com/app/ServerAppApplicationTests.java
@@ -1,13 +1,0 @@
-package com.app;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ServerAppApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
