+++ server-app/src/test/java/demo/dc/ServerAppApplicationTests.java
@@ -1,0 +1,13 @@
+package demo.dc;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ServerAppApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
